requirements.txt
pandas
pillow
numpy
pytorch-metric-learning
